+++ src/pretalx/orga/signals.py
@@ -70,6 +70,28 @@ Additionally, the signal will be called with the ``request`` it is processing.
 The receivers are expected to return HTML.
 """
 
+plugin_forms = EventPluginSignal()
+"""
+This signal allows you to put additional form elements inside the forms on several pages
+in the organiser backend. You will get the request as the keyword argument ``request``
+and are expected to return plain HTML.
+
+As with all plugin signals, the ``sender`` keyword argument will contain the event.
+Additionally, the signal will be called with the ``request`` it is processing.
+The receivers are expected to return one or more forms.
+"""
+
+plugin_html = EventPluginSignal()
+"""
+This signal allows you to put code inside the HTML body of several pages in the
+organiser backend. You will get the request as the keyword argument ``request`` and are
+expected to return plain HTML.
+
+As with all plugin signals, the ``sender`` keyword argument will contain the event.
+Additionally, the signal will be called with the ``request`` it is processing.
+The receivers are expected to return HTML.
+"""
+
 event_copy_data = EventPluginSignal()
 """
 This signal is sent out when a new event is created as a clone of an existing event, i.e.